Move the entities in their seperate folders
